// src/app/models/usuario.ts
export class Usuario{
    nombre: String;
    apellidopaterno: String;
    apellidomaterno: String;
    materia: String;
    grupo: String;
    correo: String;
//  constructor(nombre, apellidopaterno, apellidomaterno, materia, grupo, correo){
//      this.nombre= nombre
//      this.apellidopaterno= apellidopaterno;
//      this.apellidomaterno=apellidomaterno;
//      this.materia=materia;
//      this.grupo=grupo;
//      this.correo= correo;
//  }

}